requirements.txt
Django==5.0
djangorestframework==3.15.0
django-cors-headers==4.4.0
django-crontab==0.7.1
psycopg2-binary==2.9.9